# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OmniSight RSM",
    "description": "Ресурсно-сервисная модель (РСМ) — API v1 для учёта элементов, связей и корреляции алертов.\n\n**Integration API (ТЗ §8)** — `Resources`, `Correlation`, `Meta`: контракт для correlation engine и внешних систем мониторинга.\n\n**Admin API** — `CI`, `Relations`, справочники, Autodiscover, аудит: ведение модели через UI и администрирование.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/autodiscover/connectors": {
      "get": {
        "tags": [
          "Admin API · Autodiscover"
        ],
        "summary": "List Connectors V1",
        "operationId": "list_connectors_v1_api_v1_autodiscover_connectors_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "enabled_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Enabled Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Admin API · Autodiscover"
        ],
        "summary": "Create Connector V1",
        "operationId": "create_connector_v1_api_v1_autodiscover_connectors_post",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncConnectorCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/autodiscover/connectors/{connector_id}": {
      "patch": {
        "tags": [
          "Admin API · Autodiscover"
        ],
        "summary": "Update Connector V1",
        "operationId": "update_connector_v1_api_v1_autodiscover_connectors__connector_id__patch",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "connector_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Connector Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncConnectorUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Admin API · Autodiscover"
        ],
        "summary": "Delete Connector V1",
        "operationId": "delete_connector_v1_api_v1_autodiscover_connectors__connector_id__delete",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "connector_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Connector Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/autodiscover/connectors/{connector_id}/test": {
      "post": {
        "tags": [
          "Admin API · Autodiscover"
        ],
        "summary": "Test Connector V1",
        "operationId": "test_connector_v1_api_v1_autodiscover_connectors__connector_id__test_post",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "connector_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Connector Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/autodiscover/connectors/{connector_id}/sync": {
      "post": {
        "tags": [
          "Admin API · Autodiscover"
        ],
        "summary": "Sync Connector V1",
        "operationId": "sync_connector_v1_api_v1_autodiscover_connectors__connector_id__sync_post",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "connector_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Connector Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/autodiscover/profiles": {
      "get": {
        "tags": [
          "Admin API · Autodiscover"
        ],
        "summary": "List Profiles V1",
        "operationId": "list_profiles_v1_api_v1_autodiscover_profiles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/autodiscover/scan": {
      "post": {
        "tags": [
          "Admin API · Autodiscover"
        ],
        "summary": "Autodiscover Scan V1",
        "operationId": "autodiscover_scan_v1_api_v1_autodiscover_scan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutodiscoverScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/autodiscover/runs": {
      "get": {
        "tags": [
          "Admin API · Autodiscover"
        ],
        "summary": "List Runs V1",
        "operationId": "list_runs_v1_api_v1_autodiscover_runs_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/autodiscover/runs/{run_id}": {
      "get": {
        "tags": [
          "Admin API · Autodiscover"
        ],
        "summary": "Get Run V1",
        "operationId": "get_run_v1_api_v1_autodiscover_runs__run_id__get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/autodiscover/runs/{run_id}/apply": {
      "post": {
        "tags": [
          "Admin API · Autodiscover"
        ],
        "summary": "Apply Run V1",
        "operationId": "apply_run_v1_api_v1_autodiscover_runs__run_id__apply_post",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutodiscoverApplyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AutodiscoverApplyRequest": {
        "properties": {
          "mapping_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mapping Ids"
          },
          "apply_auto_only": {
            "type": "boolean",
            "title": "Apply Auto Only",
            "default": false
          }
        },
        "type": "object",
        "title": "AutodiscoverApplyRequest"
      },
      "AutodiscoverScanRequest": {
        "properties": {
          "profile_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Id"
          },
          "connector_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connector Ids"
          },
          "server_ci_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Server Ci Ids",
            "description": "Server CI group to scan"
          },
          "source_types": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Types",
            "description": "Connector types filter: host, api, file, db, stream"
          },
          "scope_mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "graph",
                  "filters",
                  "all"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope Mode"
          },
          "scope_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Scope Config"
          },
          "scope_depth": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 5.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope Depth"
          },
          "root_ci_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Root Ci Id"
          },
          "mapping_rules": {
            "additionalProperties": true,
            "type": "object",
            "title": "Mapping Rules"
          },
          "discover_relations": {
            "type": "boolean",
            "title": "Discover Relations",
            "default": true
          },
          "create_missing_ci": {
            "type": "boolean",
            "title": "Create Missing Ci",
            "default": true
          },
          "max_retries": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Retries",
            "default": 3
          },
          "auto_apply": {
            "type": "boolean",
            "title": "Auto Apply",
            "description": "Auto-apply all mappings except conflicts after scan",
            "default": true
          }
        },
        "type": "object",
        "title": "AutodiscoverScanRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SyncConnectorCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Name"
          },
          "connector_type": {
            "type": "string",
            "enum": [
              "api",
              "file",
              "db",
              "stream",
              "host"
            ],
            "title": "Connector Type"
          },
          "server_ci_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Server Ci Id"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          },
          "credentials": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SyncConnectorCredentials"
              },
              {
                "type": "null"
              }
            ]
          },
          "timeout_seconds": {
            "type": "integer",
            "maximum": 300.0,
            "minimum": 1.0,
            "title": "Timeout Seconds",
            "default": 30
          },
          "max_retries": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Max Retries",
            "default": 3
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "auto_sync": {
            "type": "boolean",
            "title": "Auto Sync",
            "description": "Periodically pull and apply discovered data",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "connector_type"
        ],
        "title": "SyncConnectorCreate"
      },
      "SyncConnectorCredentials": {
        "properties": {
          "auth_type": {
            "type": "string",
            "enum": [
              "none",
              "basic",
              "bearer",
              "api_key",
              "ssh_key"
            ],
            "title": "Auth Type",
            "default": "none"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "username_env": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username Env"
          },
          "password_env": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password Env"
          },
          "private_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Private Key"
          },
          "private_key_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Private Key Path"
          },
          "private_key_path_env": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Private Key Path Env"
          },
          "token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token"
          },
          "token_env": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Env"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "api_key_env": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key Env"
          },
          "api_key_header": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key Header",
            "default": "X-API-Key"
          },
          "database_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Database Url"
          },
          "database_url_env": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Database Url Env"
          }
        },
        "type": "object",
        "title": "SyncConnectorCredentials",
        "description": "Write-only credentials. Prefer *_env fields in production."
      },
      "SyncConnectorUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "server_ci_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Server Ci Id"
          },
          "config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config"
          },
          "credentials": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SyncConnectorCredentials"
              },
              {
                "type": "null"
              }
            ]
          },
          "timeout_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 300.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout Seconds"
          },
          "max_retries": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Retries"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "auto_sync": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Sync"
          }
        },
        "type": "object",
        "title": "SyncConnectorUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "JWT из `POST /api/v1/auth/token` (поле `session.access_token`). В Authorize вставьте только токен."
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Webhook-Secret",
        "description": "Секрет webhook-интеграций."
      },
      "XApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Ключ интеграции для correlation ingest и внешних систем."
      }
    }
  },
...
}